+++ .travis/native.sh
@@ -82,12 +82,14 @@ fi
      -O run -q --assert-output-fact 1x3xf32
 
 ./target/release/tract $CACHEDIR/hey_snips_v4_model17.pb \
-     -i Sx20xf32 --pulse 8 cost -q --assert-cost "FMA(F32)=2060448,Div(F32)=24576"
+     -i Sx20xf32 --pulse 8 cost -q \
+     --assert-cost "FMA(F32)=2060448,Div(F32)=24576,Buffer(F32)=2920"
 
 ./target/release/tract $CACHEDIR/librispeech_clean_tdnn_lstm_1e_256/model.raw \
     -f kaldi --output-node output \
     --kaldi-downsample 3 --kaldi-left-context 5 --kaldi-right-context 15 --kaldi-adjust-final-offset -5 \
-    -i Sx40 --pulse 24 cost -q --assert-cost "FMA(F32)=23725568,Div(F32)=20480"
+    -i Sx40 --pulse 24 cost -q \
+    --assert-cost "FMA(F32)=23725568,Div(F32)=20480,Buffer(F32)=1896"
 
 [ -e kaldi/test_cases/librispeech_clean_tdnn_lstm_1e_256 ] \
     || ln -s $CACHEDIR/librispeech_clean_tdnn_lstm_1e_256 kaldi/test_cases/