+++ cli/src/cost.rs
@@ -12,6 +12,7 @@ fn parse_costs(spec: &str) -> TVec<(Cost, usize)> {
             let c = match name {
                 "FMA(F32)" => Cost::FMA(f32::datum_type()),
                 "Div(F32)" => Cost::Div(f32::datum_type()),
+                "Buffer(F32)" => Cost::Buffer(f32::datum_type()),
                 _ => panic!("Unknown cost specifier {}", name),
             };
             (c, n)